Added ID Generator
Changed the previous id generation method, introduced a new class Idgen
that will be responsible for the allocation and release of Ids, while
passing the responsibility to VirtHCK class to assign an id on creation
and release it when it finishes to run or fails.
Idgen also contains a garbage collector which will kill autohck in case
it times out. timeout is configured in config.json file in units of Days.

Signed-off-by: Basil Salman <basil@daynix.com>