Updating return_when_client_up mechanism to use Thread's built in timeout
return_when_client_up times out after CLIENT_COOLDOWN_TIMEOUT seconds,
which indicates that the client could not be recognized, it was
implemented using Ruby's Timeout class, changed it to use the built
in Thread timeout on join method.
And added an aditional CLIENT_COOLDOWN_SLEEP to the configure_machine.

Signed-off-by: Bishara AbuHattoum <bishara@daynix.com>