Fixing weird behavior uppon target being of type TargetCollection
When running tests that include the action of removing a device on a
target of type TargetCollection can lead to a change in the target's
key, which in turn leads to failure in listing tests.

The fix is to verify if a target's key changed and if so, update the
target and reconnect the tools.

Signed-off-by: Bishara AbuHattoum <bishara@daynix.com>