Add QemuHCK setup manager
QemuHCK is a port of VirtHCK to ruby. QemuHCK contains all
features, which are used by AutoHCK, but not all VirtHCK
features.

For example, macvtap network support is not implemented
in the new setup manager, because AutoHCK does not use
it at all.

In the future, all missing features can be implemented.

Signed-off-by: Kostiantyn Kostiuk <konstantin@daynix.com>