QemuHCK: Let studio manage network control bridge
Before this change, QemuMachine used to run the command to delete the
network control bridge for each client. The later command executions was
failing because the bridge was already removed with the earlier command
and it did no longer exist.

With this change, the studio machine solely manages the network control
bridge, and clients do no longer try to delete the network control
bridge by its own. In fact, such a behavior was already present for the
other network devices, and this change merely expands the scope of the
behavior.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>