Let the respective owners clean threads
The trap handler used to clean threads, but the owner of the theads may
want to clean them with its own preferred timing.

Let the respective owners clean threads, and allow proper ordering of
cleanups. The change saved a few lines of AutoHCK::QemuMachine.

The thread creation in AutoHCK::Trap.normal_trap is removed instead of
adding a proper cleanup since the usage of thread can be substituted
with simple timestamps.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>