auto_hck: Always execute @project.close by its own
There used to be redundant statements of @project.close. The redundancy
was not good and one of the duplicate statements was even in a seperate
module AutoHCK::Trap, which makes the code base hard to understand.

The statement in AutoHCK::Trap may even not work because it calls
Mutex#synchronize, which is unsafe to call from a trap handler,
according to: https://docs.ruby-lang.org/en/3.1/signals_rdoc.html
The method used to be called in the following chain:
1. AutoHCK::Trap.perform_trap
2. AutoHCK::Project#close
3. AutoHCK::Engine#close
4. AutoHCK::HCKTest#close
5. AutoHCK::HCKStudio#abort
6.  AutoHCK::Tools#close
7. AutoHCK::Tools::ThreadSafe#method_missing

The duplicate statements were introduced to make sure that the code is
always executed, whether an exception or a trap raised. Fortunately
Ruby provides such a functionality with ensure clause, which this change
exploits to get rid of the redundancy and fix the described problems.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>