Create a new process group when spawning
When you interrupt AutoHCK via a shell, it signals the process group.
However, AutoHCK may want to handle the signals for the child processes
in a different way. For example, AutoHCK wants to trap supurious
SIGINTs. Create seperate process groups for them and prevent signals
for the parent process group from being propagated.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>