Revert "tests: Do not sleep after applying filters"
According to current MS examples:
https://docs.microsoft.com/en-us/windows-hardware/test/hlk/developer/filtering-code-sample

apply filters to the test result. Will try to find matching
filters for the failing task results in the project.
It is synchronous (or blocking) call, and depending on the
size of the log files and the number of task results
to be filtered in the project, the Filter method could take a while.
filterEngine.Filter(testResult);

The toolsHCK apply_project_filters uses the same API but list_tests
still returns old results for some time.

So revert commit to add sleep after applying filters
to get accurate results.

This reverts commit a513bf3397241dd54dfa91dbb970b5a5f6bd79a0.

Signed-off-by: Kostiantyn Kostiuk <konstantin@daynix.com>