Divide AutoHCK::SetuoManager#run
Before this change, studio or client was chosen according to the name
parameter. However, the values of the name parameter were not defined
in one place. Particularly, the name of studio machine was defined in
two places, AutoHCK::HCKStudio::STUDIO and AutoHCK::QemuHCK::STUDIO.
Eliminate this potential hazard by introducing specialized methods for
running studio or client machines.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>