Immediately run when HCK is instantiated
It is a common idiom in Ruby to perform operating system calls during
object initialization to implement the
Resource-Acquisition-Is-Initialization pattern and to gain its benefit.
One of such an existing example is File class, which performs file
open operation during initialization; that clarifies the close method
must be called when an instance exists.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>