Introduce AutoHCK::ResourceScope
Before this change, the management of resources that require explicit
deallocations were done ad-hoc and inconsistent. This new class
provides a unified mechanism for this purpose and ensures the following:
- Resource deallocation happens in reverse-allocation order.
  This is important because an object often depend on another object
  allocated earlier and cannot outlive the earlier object.
- Interrupts during deallocation will be postponed.
  (Ensured with Thread.handle_interrupt(AutoHCKInterrupt => :never))
- An interrupt won't be handled while an allocated object is assigned to
  a AutoHCK::ResourceScope.
  (Ensured with
   Thread.handle_interrupt(AutoHCKInterrupt => :on_blocking))

For Thread.handle_interrupt invocation, a dedicated exception class,
AutoHCKInterrupt is used so that it will not interfer with other
interrupts that can happen in child threads.

Moreover, this class provides transaction feature that facilitates error
handling during resource allocation.  If an error happens in a
transaction, resources assigned to the transaction scope will be safely
deallocated. Otherwise, they will escape the transaction scope and be
available to the outer scope.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>