Use unshare's --map-user option
bin/ns had the following comment:
> util-linux 2.36 ships handy unshare command with --map-user option,
> but unfortunately it is not available on some supported systems so
> implement the feature by ourselves. --map-root-user option cannot be
> an alternative because it conflicts with newgidmap.

However, it turned out the implementation of bin/ns does not work
since Ruby 3.3.0 because it introduced code to create a thread, which
disables CLONE_NEWUSER, at initialization time.

The implementation in Ruby was necessary to support Ubuntu 20.04, which
ships util-linux 2.34. It is quite old today so abandon it and use
unshare's --map-user option to fix incompatibility with new Ruby
versions.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>