Correct Tools closure timing
HCKStudio keeps the reference to the Tools instance it creates so the
referenced instance needs to be alive until HCKStudio dies. However,
HCKStudio used to register a Tools instance to the ResourceScope when
the instance gets created, which is later than the HCKStudio
initialization. This implies the Tools instance gets closed earlier
than HCKStudio, and HCKStudio will keep the dangling reference to the
Tools instance.

This caused an exception in case
HCKTest::configure_setup_and_synchronize got interrupted. Such an
interrupt first triggers the Tools instance closure earlier than the
closure of the HCKStudio and HCKClient instances, and made HCKClient
instances, which tries to use the Tools instance, fail.

Closing the Tools instance when HCKStudio dies extends its lifetime
properly.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>