project: Remove setupmanager from workspace path
config.json used to have setupmanager names, but they are now moved to
the platforms with commit f7d1ae6. Unfortunately we forgot to update the
init_workspace method and it kept referring to config.json.

While we can rework the method to refer to a platform, having a
setupmanager name in the workspace does no longer make sense anyway
since the platform name is already included in the workspace path, and
the platform name uniquely identifies the setupmanager name as well.
Just drop it.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>