Attach control bridge to host
The flush test requires sending UDP packets to a PDU device.
We will emulate this device on the host, so we need a connection
between client VMs and the host.

Attaching the br_world to the client is bad because a client
can download updates/drivers/applications from the Internet
and spend CPU/RAM resources in vain.

Signed-off-by: Kostiantyn Kostiuk <konstantin@daynix.com>