project: Load timestamp from workspace_path
AutoHCK reruns itself in the namespace. This causes the project
to be initialized twice with different timestamps. As a result,
timestamps in the workspace and result uploaders are different
and it is complicated to match logs.

Currently, workspace_path is for internal use only. So we guarantee
that workspace_path contains a timestamp.

Signed-off-by: Kostiantyn Kostiuk <konstantin@daynix.com>